Flask_API/requirements.txt
flask
flask_restful
flask_cors
psycopg2-binary